Switch hash to 64-bit xxHash3 for deduplication
Note: this also changes the the HashMap's internal hasher to xxHash3 for
performance. Because the keys are already hashed, we could use those
values directly, but I've used xxHash3 for simplicity. Pulling in a new
dependency like identity-hash or even creating our own dummy Hasher impl
seemed a little complex..