Fix range index errors on writing (#75)
Not really related to my other PR #74 , but I ran in to a bug where a
panic would occur on finalization - when the calculated leaf directory
length was longer than the last section, the call to
`self.entries.drain(0..len)` would panic, because the len would be
larger than the number of entries left in `self.entries`.

This patch fixes this panic by running `Directory::from_entries` instead
of draining and iterating the last leaf.

I originally thought of clamping len to the minimum of `leaf.length` and
`self.entries.len()`, but this seems like it might be a bit more
understandable? Let me know if the minimum route is better.

---------

Co-authored-by: Yuri Astrakhan <yuriastrakhan@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>