FIX: losing lat/lon precision (10e-7)
Upstream uses a decimal/integer i32 representation, scaled to 10e-7,
just like OSM's native coordinate representation.

We were using f32, which is not precise enough to represent every
number in that range - f64 is. See the test differences in this commit
for examples of the precision loss.

And alternative approach could adopt an i32 backed decimal type. In
theory this would make for more efficient representation, but as this is
only some fields in the _header_ (as opposed to say, feature
geometries which have *lots* of coordinates), saving a handful of bytes
doesn't seem worth the ceremony.